moved middleware to seperate directory
