+++ server/middleware/AuthenticationMiddleware.js
