+++ server/routes/ChallengeRoutes.js
@@ -3,7 +3,7 @@ const Challenge = require("../controllers/ChallengeController");
 const router = require("express").Router();
 
 const checkRole = require('./middleware/checkRole');
-const { AsyncWrapController } = require("./ErrorHandlingMiddleware");
+const { AsyncWrapController } = require("../middleware/ErrorHandlingMiddleware");
 
 AsyncWrapController(Challenge);
