+++ server/routes/SolutionRoutes.js
@@ -1,4 +1,4 @@
-const { AsyncWrapController } = require("../routes/ErrorHandlingMiddleware");
+const { AsyncWrapController } = require("../middleware/ErrorHandlingMiddleware");
 
 const Solution = require("../controllers/SolutionController");
 AsyncWrapController(Solution);
